#29 ability to track and send to webhooks

+++ src/helpers/commando/events/ready.js
@@ -9,7 +9,7 @@ module.exports = (client) => {
 
 	client.query.selectAllQuery('humans', 'enabled', 1).then((humans) => {
 		humans.forEach((human) => {
-			if (!client.channels.keyArray().includes(human.id) && !client.users.keyArray().includes(human.id)) {
+			if (!client.channels.keyArray().includes(human.id) && !client.users.keyArray().includes(human.id) && !human.id.match(client.hookRegex)) {
 				client.log.log({ level: 'debug', message: `unregistered ${human.name} due to 404`, event: 'discord:registerCheck' })
 
 				client.query.deleteQuery('egg', 'id', human.id)
